docling_ingestion/pyproject.toml
[project]
name = "docling-ingestion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "docling>=2.0",
    "sentence-transformers>=3.0",
    "weaviate-client>=4.9",
    "boto3>=1.35",
]
